refactor: extract hyperliquid cache service
