+++ packages/kit-bg/src/apis/BackgroundApiProxy.ts
@@ -41,6 +41,7 @@ import type ServiceHardware from '../services/ServiceHardware';
 import type ServiceHardwareUI from '../services/ServiceHardwareUI';
 import type ServiceHistory from '../services/ServiceHistory';
 import type ServiceHyperliquid from '../services/ServiceHyperLiquid/ServiceHyperliquid';
+import type ServiceHyperliquidCache from '../services/ServiceHyperLiquid/ServiceHyperliquidCache';
 import type ServiceHyperliquidExchange from '../services/ServiceHyperLiquid/ServiceHyperliquidExchange';
 import type ServiceHyperliquidReferral from '../services/ServiceHyperLiquid/ServiceHyperliquidReferral';
 import type ServiceHyperliquidSubscription from '../services/ServiceHyperLiquid/ServiceHyperliquidSubscription';
@@ -369,6 +370,10 @@ class BackgroundApiProxy
   serviceHyperliquid = this._createProxyService(
     'serviceHyperliquid',
   ) as ServiceHyperliquid;
+
+  serviceHyperliquidCache = this._createProxyService(
+    'serviceHyperliquidCache',
+  ) as ServiceHyperliquidCache;
 
   serviceHyperliquidExchange = this._createProxyService(
     'serviceHyperliquidExchange',
