+++ packages/kit-bg/src/apis/IBackgroundApi.ts
@@ -41,6 +41,7 @@ import type ServiceHardware from '../services/ServiceHardware';
 import type ServiceHardwareUI from '../services/ServiceHardwareUI';
 import type ServiceHistory from '../services/ServiceHistory';
 import type ServiceHyperliquid from '../services/ServiceHyperLiquid/ServiceHyperliquid';
+import type ServiceHyperliquidCache from '../services/ServiceHyperLiquid/ServiceHyperliquidCache';
 import type ServiceHyperliquidExchange from '../services/ServiceHyperLiquid/ServiceHyperliquidExchange';
 import type ServiceHyperliquidReferral from '../services/ServiceHyperLiquid/ServiceHyperliquidReferral';
 import type ServiceHyperliquidSubscription from '../services/ServiceHyperLiquid/ServiceHyperliquidSubscription';
@@ -226,6 +227,7 @@ export interface IBackgroundApi extends IBackgroundApiBridge {
   serviceMasterPassword: ServiceMasterPassword;
   servicePrimeTransfer: ServicePrimeTransfer;
   serviceHyperliquid: ServiceHyperliquid;
+  serviceHyperliquidCache: ServiceHyperliquidCache;
   serviceHyperliquidExchange: ServiceHyperliquidExchange;
   serviceHyperliquidReferral: ServiceHyperliquidReferral;
   serviceHyperliquidWallet: ServiceHyperliquidWallet;
