+++ packages/kit-bg/src/services/ServiceHyperLiquid/ServiceHyperliquidSubscription.ts
@@ -22,10 +22,6 @@ import platformEnv from '@onekeyhq/shared/src/platformEnv';
 import accountUtils from '@onekeyhq/shared/src/utils/accountUtils';
 import { isAppVisible } from '@onekeyhq/shared/src/utils/appVisibility';
 import {
-  getPerpsL2BookSnapshotCacheKeys,
-  swrCacheUtils,
-} from '@onekeyhq/shared/src/utils/swrCacheUtils';
-import {
   clearTrackedInterval,
   trackedSetInterval,
 } from '@onekeyhq/shared/src/utils/timerRegistry';
@@ -131,44 +127,6 @@ interface ISubscriptionUpdateParams {
 interface IRequiredSubscriptionInfo {
   requiredSubSpecsMap: Record<string, ISubscriptionSpec<ESubscriptionType>>;
   params: ISubscriptionState;
-}
-
-const ALL_DEXS_ASSET_CTXS_CACHE_WRITE_INTERVAL_MS =
-  timerUtils.getTimeDurationMs({
-    seconds: 30,
-  });
-
-const L2_BOOK_SNAPSHOT_CACHE_WRITE_INTERVAL_MS = timerUtils.getTimeDurationMs({
-  seconds: 30,
-});
-
-function setL2BookSnapshotSwrCache({
-  coin,
-  nSigFigs,
-  mantissa,
-  data,
-}: {
-  coin: string;
-  nSigFigs?: number | null;
-  mantissa?: number | null;
-  data: IBook;
-}): boolean {
-  try {
-    const keys = getPerpsL2BookSnapshotCacheKeys({
-      coin,
-      nSigFigs,
-      mantissa,
-    });
-    keys.forEach((key) => swrCacheUtils.set(key, data));
-    swrCacheUtils.flushNow();
-    return true;
-  } catch (error) {
-    defaultLogger.perp.hyperliquid.cacheSnapshotError({
-      type: 'l2_book_swr',
-      error,
-    });
-    return false;
-  }
 }
 
 @backgroundClass()
@@ -964,7 +922,7 @@ export default class ServiceHyperliquidSubscription extends ServiceBase {
 
   @backgroundMethod()
   async disconnect(): Promise<void> {
-    this._flushPendingL2BookSnapshotCache();
+    this.backgroundApi.serviceHyperliquidCache.flushPendingL2BookSnapshotCache();
     this._subscriptionLifecycleVersion += 1;
     this._updateSubscriptionsDebounced.cancel();
     this._unwatchSubscriptionAtoms();
@@ -990,7 +948,7 @@ export default class ServiceHyperliquidSubscription extends ServiceBase {
 
   @backgroundMethod()
   async cleanup(): Promise<void> {
-    this._flushPendingL2BookSnapshotCache();
+    this.backgroundApi.serviceHyperliquidCache.flushPendingL2BookSnapshotCache();
     this._subscriptionLifecycleVersion += 1;
     this._updateSubscriptionsDebounced.cancel();
     this._unwatchSubscriptionAtoms();
@@ -1003,7 +961,7 @@ export default class ServiceHyperliquidSubscription extends ServiceBase {
   // Skip per-subscription unsubscribe to avoid async race where stale
   // _destroySubscription completion deletes newly created tracking entries
   private async _forceReconnectTransport(): Promise<void> {
-    this._flushPendingL2BookSnapshotCache();
+    this.backgroundApi.serviceHyperliquidCache.flushPendingL2BookSnapshotCache();
     this._subscriptionLifecycleVersion += 1;
     this._updateSubscriptionsDebounced.cancel();
     this._unwatchSubscriptionAtoms();
@@ -1950,13 +1908,22 @@ export default class ServiceHyperliquidSubscription extends ServiceBase {
           data as IWsOpenOrders,
         );
       } else if (subscriptionType === ESubscriptionType.ALL_DEXS_ASSET_CTXS) {
-        this._cacheAllDexsAssetCtxsSnapshot(data as IWsAllDexsAssetCtxs);
+        this.backgroundApi.serviceHyperliquidCache.cacheAllDexsAssetCtxsSnapshot(
+          data as IWsAllDexsAssetCtxs,
+        );
         this._emitHyperliquidDataUpdate(
           subscriptionType,
           data as IWsAllDexsAssetCtxs,
         );
       } else if (subscriptionType === ESubscriptionType.L2_BOOK) {
-        this._cacheL2BookSnapshot(data as IBook);
+        this.backgroundApi.serviceHyperliquidCache.cacheL2BookSnapshot({
+          data: data as IBook,
+          activeBookCoin:
+            this._currentState.tradingMode === 'spot'
+              ? this._currentState.currentSpotSymbol
+              : this._currentState.currentSymbol,
+          activeOptions: this._currentState.l2BookOptions,
+        });
         this._emitHyperliquidDataUpdate(subscriptionType, data);
       } else {
         this._emitHyperliquidDataUpdate(subscriptionType, data);
@@ -1986,134 +1953,6 @@ export default class ServiceHyperliquidSubscription extends ServiceBase {
   }
 
   private _lastLivenessAtomUpdate = 0;
-
-  private _lastAllDexsAssetCtxsCacheWriteAt = 0;
-
-  private _lastL2BookSnapshotCacheWriteAt = 0;
-
-  private _l2BookSnapshotCacheTimer: ReturnType<typeof setTimeout> | null =
-    null;
-
-  private _pendingL2BookSnapshotCache:
-    | {
-        coin: string;
-        nSigFigs?: number | null;
-        mantissa?: number | null;
-        data: IBook;
-      }
-    | undefined;
-
-  private _cacheAllDexsAssetCtxsSnapshot(data: IWsAllDexsAssetCtxs) {
-    const now = Date.now();
-    if (
-      now - this._lastAllDexsAssetCtxsCacheWriteAt <
-      ALL_DEXS_ASSET_CTXS_CACHE_WRITE_INTERVAL_MS
-    ) {
-      return;
-    }
-
-    const ctxCount =
-      data?.ctxs?.reduce((sum, [, ctxs]) => sum + (ctxs?.length ?? 0), 0) ?? 0;
-    if (ctxCount <= 0) {
-      return;
-    }
-
-    this._lastAllDexsAssetCtxsCacheWriteAt = now;
-    markPerpsColdStartPerfOnce(
-      'service_all_dexs_asset_ctxs_cache_write_first',
-      {
-        dexCount: data.ctxs.length,
-        ctxCount,
-      },
-    );
-    void this.backgroundApi.simpleDb.perp
-      .setAllDexsAssetCtxsSnapshotCache(data)
-      .catch((error) => {
-        defaultLogger.perp.hyperliquid.cacheSnapshotError({
-          type: 'all_dexs_asset_ctxs_simple_db',
-          error,
-        });
-      });
-  }
-
-  private _writeL2BookSnapshotCache(
-    payload: NonNullable<typeof this._pendingL2BookSnapshotCache>,
-  ) {
-    this._lastL2BookSnapshotCacheWriteAt = Date.now();
-    const didWriteSwrCache = setL2BookSnapshotSwrCache(payload);
-    markPerpsColdStartPerfOnce('service_l2_book_ws_cache_write_first', {
-      coin: payload.coin,
-      bidLevels: payload.data.levels?.[0]?.length ?? 0,
-      askLevels: payload.data.levels?.[1]?.length ?? 0,
-      nSigFigs: payload.nSigFigs,
-      mantissa: payload.mantissa,
-      swr: didWriteSwrCache,
-    });
-    void this.backgroundApi.simpleDb.perp
-      .setL2BookSnapshotCache(payload)
-      .catch((error) => {
-        defaultLogger.perp.hyperliquid.cacheSnapshotError({
-          type: 'l2_book_simple_db',
-          error,
-        });
-      });
-  }
-
-  private _flushPendingL2BookSnapshotCache() {
-    if (this._l2BookSnapshotCacheTimer) {
-      clearTimeout(this._l2BookSnapshotCacheTimer);
-      this._l2BookSnapshotCacheTimer = null;
-    }
-    const pending = this._pendingL2BookSnapshotCache;
-    this._pendingL2BookSnapshotCache = undefined;
-    if (pending) {
-      this._writeL2BookSnapshotCache(pending);
-    }
-  }
-
-  private _cacheL2BookSnapshot(data: IBook) {
-    const coin = data?.coin;
-    const hasLevels = Boolean(
-      data?.levels?.[0]?.length && data?.levels?.[1]?.length,
-    );
-    if (!coin || !hasLevels) {
-      return;
-    }
-
-    const activeOptions = this._currentState.l2BookOptions;
-    const activeBookCoin =
-      this._currentState.tradingMode === 'spot'
-        ? this._currentState.currentSpotSymbol
-        : this._currentState.currentSymbol;
-    const isActiveBook = activeBookCoin === coin;
-    const payload = {
-      coin,
-      nSigFigs: isActiveBook ? (activeOptions?.nSigFigs ?? null) : null,
-      mantissa: isActiveBook ? (activeOptions?.mantissa ?? null) : null,
-      data,
-    };
-    const now = Date.now();
-    const elapsed = now - this._lastL2BookSnapshotCacheWriteAt;
-    if (elapsed >= L2_BOOK_SNAPSHOT_CACHE_WRITE_INTERVAL_MS) {
-      this._pendingL2BookSnapshotCache = undefined;
-      this._writeL2BookSnapshotCache(payload);
-      return;
-    }
-
-    this._pendingL2BookSnapshotCache = payload;
-    if (this._l2BookSnapshotCacheTimer) {
-      return;
-    }
-
-    this._l2BookSnapshotCacheTimer = setTimeout(() => {
-      this._l2BookSnapshotCacheTimer = null;
-      const pending = this._pendingL2BookSnapshotCache;
-      this._pendingL2BookSnapshotCache = undefined;
-      if (pending) {
-        this._writeL2BookSnapshotCache(pending);
-      }
-    }, L2_BOOK_SNAPSHOT_CACHE_WRITE_INTERVAL_MS - elapsed);
-  }
 
   private _updateNetworkLiveness() {
     const now = Date.now();
